Create product price endpoint
This patch adds the endpoint to check a product price in all the
exchanges available to check in the application, which are defined by
the environment variable.

This is done by adding the endpoint in the blueprint of views and by
extending the service in order to use another resource in moneeda api.