Implement InvalidProduct exception
Since the new product prices flow takes product id as an input, the
server needs to let the user know if the requested product doesn't exist
in the first place.

This patch implements an exception in the service moneeda that is raised
if the product passed in the route is invalid, returning a BadRequest to
the user and a JSON with a message.