Initial setup of the SPA frontend
This commit is intended to setup the boilerplate code to use webpack,
react and redux in order to do a simple Single Page Application (SPA) to
check the products available in the exchanges that the backend is ready
to serve.

In this patch, there is only one action intended to filter something in
the input box, with some pre coded product examples. This is the only
action possible, which was used to setup the boilerplate code to redux.

In future commits, there will be more complicated actions, to deal with
client-server communication and adjusts in the layout.