Create flux to fetch products from backend node
In this commit, the redux flux using the action reducers to fetch the
products in the backend is implemented.

In this purpose, there was need to implement an API object to handle
client-server communication, a XHR API object and all the code around
the redux.

Also, due to the async nature of this task, the redux-thunk middleware
was introduced in the package json.

The UI is not user friendly yet, but this will be changed in future
patches.